refactor : Refactor string calculator
