+++ src/main/java/car/racing/Car.java
@@ -1,11 +1,15 @@
 package car.racing;
 
+import java.util.Random;
+
 public class Car {
 
     private static final String FORWARD = "-";
+    private static final int LIMIT = 10;
 
     private final Forwardable forwardable;
     private final StringBuilder forwardStrBuilder = new StringBuilder();
+    private final Random random = new Random();
 
     public Car(Forwardable forwardable) {
         this.forwardable = forwardable;
@@ -16,7 +20,7 @@ public class Car {
     }
 
     public void forwardOrNot() {
-        if (forwardable.forwardable()) {
+        if (forwardable.forwardable(random.nextInt(LIMIT))) {
             forwardStrBuilder.append(FORWARD);
         }
     }
