+++ src/main/java/car/racing/CarRacingMain.java
@@ -4,6 +4,6 @@ public class CarRacingMain {
 
     public static void main(String[] args) {
         InputView inputView = new InputView();
-        inputView.execute(CarRacingControllerFactory.newInstance());
+        inputView.execute(ViewControllerFactory.newInstance());
     }
 }
