+++ src/main/java/car/racing/CarRacingManager.java
@@ -6,12 +6,13 @@ import java.util.stream.IntStream;
 
 public class CarRacingManager {
 
+    private final Forwardable forwardable = new ForwardableImpl();
     final List<Car> cars = new ArrayList<>();
 
     public void setupCars(int numOfCar) {
         cars.clear();
         IntStream.range(0, numOfCar)
-                .forEach(i -> cars.add(new Car(new ForwardableImpl())));
+                .forEach(i -> cars.add(new Car(forwardable)));
     }
 
     public List<Car> forwardEachCar() {
