+++ src/main/java/car/racing/ForwardableImpl.java
@@ -1,16 +1,11 @@
 package car.racing;
 
-import java.util.Random;
-
 public class ForwardableImpl implements Forwardable {
 
-    private static final int LIMIT = 10;
     private static final int THRESHOLD = 4;
 
-    private final Random random = new Random();
-
     @Override
-    public boolean forwardable() {
-        return random.nextInt(LIMIT) >= THRESHOLD;
+    public boolean forwardable(int forward) {
+        return forward >= THRESHOLD;
     }
 }
