+++ src/main/java/string/calculator/CalculatorBinaryOperator.java
@@ -1,4 +1,4 @@
-package string_calculator;
+package string.calculator;
 
 public interface CalculatorBinaryOperator {
