+++ src/main/java/string/calculator/StringCalculator.java
@@ -1,4 +1,4 @@
-package string_calculator;
+package string.calculator;
 
 
 import java.util.LinkedList;
