+++ src/main/java/string/calculator/StringUtils.java
@@ -1,4 +1,4 @@
-package string_calculator;
+package string.calculator;
 
 public class StringUtils {
